pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.97.1",
    "pathlib2>=2.3.7.post1",
    "playwright>=1.54.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "robotframework>=7.3.2",
    "werkzeug>=3.1.3",
    "flask-migrate>=4.1.0",
    "python-dotenv>=1.1.1",
    "robotframework-browser>=19.6.2",
]
